chatguru-clickup-middleware/cloud_functions/vertex_media_processor/requirements.txt
google-cloud-pubsub==2.23.1
google-cloud-aiplatform==1.71.1
requests==2.32.3
